🌱 [scanner] refactor: split registerHooks and WidgetExportModalPreview
Split web/src/lib/unified/registerHooks.ts (927 → 44 LOC):
- Extract factory helpers to registerHooks/factories.ts
- Extract hook config tables to registerHooks/hookTables.ts
- Extract demo data constants to registerHooks/demoData.ts
- Wire up existing registerHooks/demoSupport.ts and customHooks.ts

Split web/src/components/widgets/widget-export-modal/WidgetExportModalPreview.tsx (939 → 178 LOC):
- Extract shared style constants to previewStyles.ts
- Extract first card preview group to CardPreviewA.tsx
- Extract second card preview group + NightlyE2EPreview to CardPreviewB.tsx
- Main file becomes a thin router/dispatcher

Part of #21375 — oversized source files (auto-qa threshold: 500 LOC)
