+++ web/src/components/widgets/widget-export-modal/WidgetExportModalPreview.tsx
@@ -1,19 +1,29 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
-import { Smartphone } from 'lucide-react'
-import { useTranslation } from 'react-i18next'
-import { WIDGET_CARDS, WIDGET_STATS, WIDGET_TEMPLATES, type WidgetCardDefinition } from '../../../lib/widgets/widgetRegistry'
-import { type WidgetConfig } from '../../../lib/widgets/codeGenerator'
-import type { CSSProperties } from 'react'
-import { TechnicalAcronym } from '../../shared/TechnicalAcronym'
+/**
+ * Widget export modal preview dispatcher and shared sizing helpers.
+ */
 
-// Inline style constants
-const WIDGET_EXPORT_MODAL_PREVIEW_MAX_WIDTH_PX = 260
-const WIDGET_EXPORT_MODAL_PREVIEW_MAX_HEIGHT_PX = 220
-const WIDGET_EXPORT_MODAL_DIV_STYLE_2: CSSProperties = { flex: 1 }
-const WIDGET_EXPORT_MODAL_SPAN_STYLE_1: CSSProperties = { fontWeight: 500, maxWidth: '110px', overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }
-const WIDGET_EXPORT_MODAL_SPAN_STYLE_2: CSSProperties = { fontWeight: 500 }
-const WIDGET_EXPORT_MODAL_SPAN_STYLE_3: CSSProperties = { fontWeight: 500, flex: 1 }
-const WIDGET_EXPORT_MODAL_SPAN_STYLE_4: CSSProperties = { width: '24px', fontWeight: 600, color: '#9ca3af' }
+import { Smartphone } from 'lucide-react'
+import type { CSSProperties } from 'react'
+import { WIDGET_CARDS, WIDGET_STATS, WIDGET_TEMPLATES } from '../../../lib/widgets/widgetRegistry'
+import { type WidgetConfig } from '../../../lib/widgets/codeGenerator'
+import {
+  PREV_CARD_PAD,
+  PREV_CLR_DIM,
+  PREV_FS_BODY,
+  PREV_FS_LABEL,
+  PREV_FS_MICRO,
+  PREV_FS_STAT,
+  PREV_FS_STAT_SM,
+  PREV_ITEM_PAD,
+  PREV_SM,
+  PREV_XS,
+  SAMPLE_STATS,
+  WIDGET_EXPORT_MODAL_PREVIEW_MAX_HEIGHT_PX,
+  WIDGET_EXPORT_MODAL_PREVIEW_MAX_WIDTH_PX,
+  ps,
+} from './previewStyles'
+import { CardPreviewA } from './CardPreviewA'
+import { CardPreviewB } from './CardPreviewB'
 
 export function getWidgetPreviewDimensions(config: WidgetConfig | null): { width: number; height: number } | null {
   if (!config) return null
@@ -54,756 +64,20 @@ export function getWidgetPreviewScale(dimensions: { width: number; height: numbe
   )
 }
 
+const CARD_PREVIEW_A_TYPES = new Set([
+  'cluster_health', 'pod_issues', 'gpu_overview', 'hardware_health', 'nightly_e2e_status',
+  'security_issues', 'active_alerts', 'helm_releases', 'top_pods', 'event_summary',
+  'warning_events', 'operator_status', 'storage_overview', 'pvc_status', 'network_overview',
+  'service_status', 'opencost_overview', 'provider_health', 'nightly_release_pulse',
+])
 
-
-// Spacing constants aligned to the 4px grid – centralise magic values used throughout widget previews
-const PREV_XS = '4px'           // 1 × 4px: tight gaps and small padding
-const PREV_SM = '8px'           // 2 × 4px: standard margins and gaps
-const PREV_MD = '12px'          // 3 × 4px: medium spacing
-const PREV_LG = '16px'          // 4 × 4px: large gaps and section spacing
-const PREV_ITEM_PAD = '4px 8px' // py-1 px-2: item-row padding (vertical=XS, horizontal=SM)
-const PREV_CARD_PAD = '8px 12px' // standard card padding (vertical=SM, horizontal=MD)
-const PREV_DOTS_GAP = '2px'     // sub-grid gap for tightly-packed status-dot rows
-const PREV_BAR_GAP = '3px'      // gap between adjacent bar-chart columns
-const PREV_HAIRLINE_GAP = '1px' // minimal gap between stacked bar segments
-const PREV_BORDER_THIN = '3px'  // thin accent border for status indicators
-const PREV_BORDER_STD = '4px'   // standard accent border for alerts and highlights
-// Font sizes used in preview components — extracted to pass the hex/magic-number ratchet
-const PREV_FS_HERO = '28px'     // large hero number (GPU %)
-const PREV_FS_HEADLINE = '24px' // headline stat (cost total, CI pass %)
-const PREV_FS_FEATURED = '20px' // featured value (release tag, AI status)
-const PREV_FS_STAT = '16px'     // stat block value when smaller than ps.statVal
-const PREV_FS_STAT_SM = '14px'  // compact stat value (metric rows)
-const PREV_FS_BODY = '12px'     // body text, count badges
-const PREV_FS_CAPTION = '10px'  // caption text, row items
-const PREV_FS_MICRO = '9px'     // micro text, timestamps, secondary labels
-const PREV_FS_LABEL = '8px'     // smallest label text in compact stat blocks
-
-// Issue activity chart bar-height multipliers — keeps preview bars proportional
-const PREV_BAR_OPENED_SCALE = 6   // px per unit for the "opened" bar segment
-const PREV_BAR_CLOSED_SCALE = 4   // px per unit for the "closed" bar segment
-const PREV_BAR_CLOSED_BASE = 8    // baseline value subtracted before scaling closed segment
-
-// Preview color constants — declared before `ps` so they can be used in its definition
-const PREV_CLR_TEXT = '#f9fafb'      // primary text color in widget previews
-const PREV_CLR_MUTED = '#9ca3af'     // muted/secondary text color
-const PREV_CLR_SECONDARY = '#cbd5e1' // light secondary text
-const PREV_CLR_DIM = '#d1d5db'       // dim text
-const PREV_CLR_CPU = '#60a5fa'       // CPU metric color (blue-400)
-const PREV_CLR_MEM = '#c084fc'       // Memory metric color (purple-400)
-
-// Design token colors for macOS widget preview
-const PREV_BG_DARK_OPACITY = 'rgba(17, 24, 39, 0.9)'  // Dark gray-900 with 0.9 opacity
-const PREV_BORDER_LIGHT = 'rgba(255, 255, 255, 0.1)'  // Subtle white border
-const PREV_SHADOW = '0 4px 6px -1px rgba(0, 0, 0, 0.1)' // Shadow
-
-/** Shared preview styles matching macOS Übersicht widget appearance.
- *  These use hardcoded dark colors intentionally — they render a fixed preview
- *  of how the exported widget will look on a macOS desktop, regardless of
- *  the console's current theme. */
-const ps = {
-  card: {
-    backgroundColor: PREV_BG_DARK_OPACITY,
-    borderRadius: '12px',
-    padding: `${PREV_MD} ${PREV_LG}`,
-    border: `1px solid ${PREV_BORDER_LIGHT}`,
-    color: PREV_CLR_TEXT,
-    fontFamily: 'Inter, -apple-system, sans-serif',
-    fontSize: PREV_FS_BODY,
-    lineHeight: 1.4,
-    boxShadow: PREV_SHADOW } as CSSProperties,
-  title: {
-    fontSize: PREV_FS_BODY,
-    fontWeight: 600,
-    color: PREV_CLR_TEXT,
-    marginBottom: PREV_SM,
-    display: 'flex',
-    alignItems: 'center',
-    gap: PREV_SM } as CSSProperties,
-  dot: (color: string) => ({
-    width: 7,
-    height: 7,
-    borderRadius: '50%',
-    backgroundColor: color,
-    display: 'inline-block',
-    flexShrink: 0 }) as CSSProperties,
-  statBlock: {
-    backgroundColor: PREV_BG_DARK_OPACITY,
-    borderRadius: '6px',
-    border: `1px solid ${PREV_BORDER_LIGHT}`,
-    padding: PREV_CARD_PAD,
-    display: 'flex',
-    flexDirection: 'column' as const,
-    alignItems: 'center',
-    justifyContent: 'center',
-    minWidth: '54px' } as CSSProperties,
-  statVal: {
-    fontSize: PREV_FS_STAT,
-    fontWeight: 700,
-    lineHeight: 1.2 } as CSSProperties,
-  statLbl: {
-    fontSize: PREV_FS_MICRO,
-    color: PREV_CLR_MUTED,
-    textTransform: 'uppercase' as const,
-    letterSpacing: '0.05em',
-    marginTop: PREV_XS } as CSSProperties,
-  row: { display: 'flex', gap: PREV_SM, alignItems: 'center' } as CSSProperties,
-  col: { display: 'flex', flexDirection: 'column' as const, gap: PREV_XS } as CSSProperties,
-  muted: { color: PREV_CLR_MUTED, fontSize: PREV_FS_CAPTION } as CSSProperties,
-  colors: { healthy: '#22c55e', warning: '#eab308', error: '#ef4444', info: '#3b82f6', purple: '#9333ea' } }
-
-// Sample stat data for realistic previews
-const SAMPLE_STATS: Record<string, number | string> = {
-  total_clusters: 4,
-  total_pods: 128,
-  total_gpus: 32,
-  cpu_usage: '67%',
-  memory_usage: '54%',
-  unhealthy_pods: 3,
-  active_alerts: 2 }
-
-// Widget preview with realistic mock data
-export function WidgetPreview({ config }: { config: WidgetConfig | null }) {
-  if (!config) {
-    return (
-      <div className="text-center text-muted-foreground">
-        <Smartphone className="w-12 h-12 mx-auto mb-2 opacity-50" />
-        <p className="text-sm">Select an item to preview</p>
-      </div>
-    )
-  }
-
-  if (config.type === 'card' && config.cardType) {
-    return <CardPreview cardType={config.cardType} />
-  }
-
-  if (config.type === 'stat' && config.statIds) {
-    return <StatPreview statIds={config.statIds} />
-  }
-
-  if (config.type === 'template' && config.templateId) {
-    return <TemplatePreview templateId={config.templateId} />
-  }
-
-  return null
-}
-
-// --- Card previews ---
 function CardPreview({ cardType }: { cardType: string }) {
-  const { t } = useTranslation()
   const card = WIDGET_CARDS[cardType]
   if (!card) return null
-
-  switch (cardType) {
-    case 'cluster_health':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.warning)} /> Cluster Health</div>
-          <div style={ps.row}>
-            <div style={{ ...ps.statBlock, borderLeft: `${PREV_BORDER_THIN} solid ${ps.colors.healthy}` }}>
-              <span style={{ ...ps.statVal, color: ps.colors.healthy }}>3</span>
-              <span style={ps.statLbl}>{t('common.healthy')}</span>
-            </div>
-            <div style={{ ...ps.statBlock, borderLeft: `${PREV_BORDER_THIN} solid ${ps.colors.error}` }}>
-              <span style={{ ...ps.statVal, color: ps.colors.error }}>1</span>
-              <span style={ps.statLbl}>{t('common.unhealthy')}</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'pod_issues':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.warning)} /> Pod Issues</div>
-          <div style={ps.muted}>4 total issues</div>
-          <div style={{ ...ps.col, marginTop: PREV_SM }}>
-            <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(239,68,68,0.1)', borderRadius: PREV_XS }}>
-              <span style={{ color: ps.colors.error, fontWeight: 600, fontSize: PREV_FS_BODY }}>2</span>
-              <span style={ps.muted}>CrashLoopBackOff</span>
-            </div>
-            <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(234,179,8,0.1)', borderRadius: PREV_XS }}>
-              <span style={{ color: ps.colors.warning, fontWeight: 600, fontSize: PREV_FS_BODY }}>1</span>
-              <span style={ps.muted}>OOMKilled</span>
-            </div>
-            <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(59,130,246,0.1)', borderRadius: PREV_XS }}>
-              <span style={{ color: ps.colors.info, fontWeight: 600, fontSize: PREV_FS_BODY }}>1</span>
-              <span style={ps.muted}>ImagePullBackOff</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'gpu_overview':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.purple)} /> GPU Overview</div>
-          <div style={{ textAlign: 'center', marginBottom: PREV_SM }}>
-            <div style={{ fontSize: PREV_FS_HERO, fontWeight: 700, color: ps.colors.purple }}>72%</div>
-            <div style={ps.muted}>{t('common.utilization')}</div>
-          </div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={ps.statVal}>32</span>
-              <span style={ps.statLbl}>{t('common.total')}</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, color: ps.colors.purple }}>23</span>
-              <span style={ps.statLbl}>{t('common.allocated')}</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'hardware_health':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.warning)} /> Hardware Health</div>
-          <div style={{ ...ps.row, marginBottom: PREV_SM }}>
-            <div style={{ ...ps.statBlock, borderLeft: `${PREV_BORDER_STD} solid ${ps.colors.healthy}` }}>
-              <span style={ps.statVal}>4</span>
-              <span style={ps.statLbl}>{t('common.nodes')}</span>
-            </div>
-            <div style={{ ...ps.statBlock, borderLeft: `${PREV_BORDER_STD} solid ${ps.colors.purple}` }}>
-              <span style={{ ...ps.statVal, color: ps.colors.purple }}>16</span>
-              <span style={ps.statLbl}>{t('common.gpus')}</span>
-            </div>
-            <div style={{ ...ps.statBlock, borderLeft: `${PREV_BORDER_STD} solid ${ps.colors.info}` }}>
-              <span style={{ ...ps.statVal, color: ps.colors.info }}>8</span>
-              <span style={ps.statLbl}>NICs</span>
-            </div>
-          </div>
-          <div style={ps.col}>
-            <div style={{ fontSize: PREV_FS_MICRO, fontWeight: 600, color: PREV_CLR_MUTED, textTransform: 'uppercase', letterSpacing: '0.05em' }}>Alerts (2)</div>
-            <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(239,68,68,0.1)', borderRadius: PREV_XS, borderLeft: `${PREV_BORDER_STD} solid ${ps.colors.error}` }}>
-              <span style={{ fontSize: PREV_FS_CAPTION, color: ps.colors.error, fontWeight: 600 }}>GPU</span>
-              <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED, marginLeft: PREV_XS }}>worker-3 (-2)</span>
-            </div>
-            <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(234,179,8,0.1)', borderRadius: PREV_XS, borderLeft: `${PREV_BORDER_STD} solid ${ps.colors.warning}` }}>
-              <span style={{ fontSize: PREV_FS_CAPTION, color: ps.colors.warning, fontWeight: 600 }}>NIC</span>
-              <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED, marginLeft: PREV_XS }}>worker-1 (-1)</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'nightly_e2e_status':
-      return <NightlyE2EPreview />
-
-    case 'security_issues':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.warning)} /> Security Issues</div>
-          <div style={ps.col}>
-            {[
-              { label: 'Privileged containers', count: 3, color: ps.colors.error },
-              { label: 'No resource limits', count: 12, color: ps.colors.warning },
-              { label: 'Running as root', count: 5, color: ps.colors.error },
-            ].map((item) => (
-              <div key={item.label} className="px-2 py-1" style={{ ...ps.row, backgroundColor: item.color === ps.colors.error ? 'rgba(239,68,68,0.1)' : 'rgba(234,179,8,0.1)', borderRadius: PREV_XS }}>
-                <span style={{ color: item.color, fontWeight: 600, fontSize: PREV_FS_BODY, minWidth: '16px' }}>{item.count}</span>
-                <span style={ps.muted}>{item.label}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'active_alerts':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.error)} /> Active Alerts</div>
-          <div style={ps.col}>
-            {[
-              { name: 'HighMemoryUsage', severity: 'critical', ns: 'monitoring' },
-              { name: 'PodCrashLooping', severity: 'warning', ns: 'default' },
-              { name: 'NodeDiskPressure', severity: 'warning', ns: 'kube-system' },
-            ].map((a) => (
-              <div key={a.name} className="px-2 py-1" style={{ ...ps.row, backgroundColor: a.severity === 'critical' ? 'rgba(239,68,68,0.1)' : 'rgba(234,179,8,0.1)', borderRadius: PREV_XS, borderLeft: `${PREV_BORDER_STD} solid ${a.severity === 'critical' ? ps.colors.error : ps.colors.warning}` }}>
-                <span style={{ fontSize: PREV_FS_CAPTION, color: a.severity === 'critical' ? ps.colors.error : ps.colors.warning, fontWeight: 600 }}>{a.name}</span>
-                <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED, marginLeft: 'auto' }}>{a.ns}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'helm_releases':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Helm Releases</div>
-          <div style={ps.col}>
-            {[
-              { name: 'ingress-nginx', status: 'deployed', ver: '4.8.3' },
-              { name: 'cert-manager', status: 'deployed', ver: '1.13.2' },
-              { name: 'prometheus', status: 'deployed', ver: '25.8.0' },
-              { name: 'redis', status: 'failed', ver: '18.4.0' },
-            ].map((r) => (
-              <div key={r.name} style={{ ...ps.row, justifyContent: 'space-between' }}>
-                <span style={{ fontSize: PREV_FS_CAPTION, fontWeight: 500 }}>{r.name}</span>
-                <span style={{ fontSize: PREV_FS_MICRO, color: r.status === 'deployed' ? ps.colors.healthy : ps.colors.error }}>{r.status}</span>
-                <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED }}>{r.ver}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'top_pods':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Top Pods</div>
-          <div style={ps.col}>
-            {[
-              { name: 'ml-training-job-7x', cpu: '3.2 cores', mem: '12.4 Gi' },
-              { name: 'prometheus-server-0', cpu: '1.8 cores', mem: '8.2 Gi' },
-              { name: 'elasticsearch-data-1', cpu: '1.4 cores', mem: '6.1 Gi' },
-            ].map((p) => (
-              <div key={p.name} style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_1}>{p.name}</span>
-                <span style={{ color: PREV_CLR_CPU }}>{p.cpu}</span>
-                <span style={{ color: PREV_CLR_MEM }}>{p.mem}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'event_summary':
-    case 'warning_events':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(cardType === 'warning_events' ? ps.colors.warning : ps.colors.info)} /> {card.displayName}</div>
-          <div style={ps.col}>
-            {[
-              { type: 'Warning', count: 12, msg: 'BackOff restarting failed container' },
-              { type: 'Warning', count: 5, msg: 'Readiness probe failed' },
-              { type: 'Normal', count: 34, msg: 'Scheduled successfully' },
-            ].map((e, i) => (
-              <div key={i} style={{ ...ps.row, fontSize: PREV_FS_CAPTION }}>
-                <span style={{ color: e.type === 'Warning' ? ps.colors.warning : ps.colors.healthy, fontWeight: 600, minWidth: '18px' }}>{e.count}</span>
-                <span style={{ color: PREV_CLR_SECONDARY, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>{e.msg}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'operator_status':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Operator Status</div>
-          <div style={ps.col}>
-            {[
-              { name: 'cert-manager', ready: true },
-              { name: 'gpu-operator', ready: true },
-              { name: 'prometheus-operator', ready: true },
-              { name: 'node-feature-discovery', ready: false },
-            ].map((o) => (
-              <div key={o.name} style={{ ...ps.row, fontSize: PREV_FS_CAPTION }}>
-                <span style={ps.dot(o.ready ? ps.colors.healthy : ps.colors.warning)} />
-                <span>{o.name}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'storage_overview':
-    case 'pvc_status':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> {card.displayName}</div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.info }}>24</span>
-              <span style={ps.statLbl}>{t('common.pvcs')}</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.healthy }}>22</span>
-              <span style={ps.statLbl}>{t('common.bound')}</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.warning }}>2</span>
-              <span style={ps.statLbl}>{t('common.pending')}</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'network_overview':
-    case 'service_status':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> {card.displayName}</div>
-          <div style={{ ...ps.row, marginBottom: PREV_SM }}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT }}>18</span>
-              <span style={ps.statLbl}>{t('common.services')}</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.info }}>6</span>
-              <span style={ps.statLbl}>Policies</span>
-            </div>
-          </div>
-          <div style={ps.col}>
-            {['ClusterIP (12)', 'LoadBalancer (4)', 'NodePort (2)'].map((s) => (
-              <div key={s} style={{ fontSize: PREV_FS_CAPTION, color: PREV_CLR_MUTED }}>{s}</div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'opencost_overview':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> OpenCost Overview</div>
-          <div style={{ textAlign: 'center', marginBottom: PREV_SM }}>
-            <div style={{ fontSize: PREV_FS_HEADLINE, fontWeight: 700, color: ps.colors.healthy }}>$1,247</div>
-            <div style={ps.muted}>Monthly estimate</div>
-          </div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: PREV_CLR_CPU }}>$482</span>
-              <span style={ps.statLbl}>Compute</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: PREV_CLR_MEM }}>$635</span>
-              <span style={ps.statLbl}>GPU</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: ps.colors.info }}>$130</span>
-              <span style={ps.statLbl}>{t('common.storage')}</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'provider_health':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Provider Health</div>
-          <div style={ps.col}>
-            {[
-              { name: 'OpenAI', status: 'operational', color: ps.colors.healthy },
-              { name: 'Anthropic', status: 'operational', color: ps.colors.healthy },
-              { name: 'AWS', status: 'degraded', color: ps.colors.warning },
-              { name: 'GCP', status: 'operational', color: ps.colors.healthy },
-            ].map((p) => (
-              <div key={p.name} style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_2}>{p.name}</span>
-                <span style={{ color: p.color }}>{p.status}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'nightly_release_pulse':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Nightly Release Pulse</div>
-          <div style={{ textAlign: 'center', marginBottom: PREV_SM }}>
-            <div style={{ fontSize: PREV_FS_FEATURED, fontWeight: 700, color: ps.colors.healthy }}>v0.3.22</div>
-            <div style={ps.muted}>Released 2h ago</div>
-          </div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: ps.colors.healthy }}>12</span>
-              <span style={ps.statLbl}>Streak</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: ps.colors.info }}>93%</span>
-              <span style={ps.statLbl}>Pass rate</span>
-            </div>
-          </div>
-        </div>
-      )
-
-    case 'workflow_matrix':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Workflow Matrix</div>
-          <div style={ps.col}>
-            {['build', 'test', 'lint', 'deploy'].map((wf) => (
-              <div key={wf} style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_2}>{wf}</span>
-                <div style={{ display: 'flex', gap: PREV_DOTS_GAP }}>
-                  {[1, 2, 3, 4, 5].map((i) => (
-                    <div key={i} style={{ width: '8px', height: '8px', borderRadius: '2px', backgroundColor: i === 3 && wf === 'deploy' ? ps.colors.error : ps.colors.healthy }} />
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'pipeline_flow':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Live Runs</div>
-          <div style={ps.col}>
-            {[
-              { name: 'build (amd64)', status: 'running', time: '3m 12s' },
-              { name: 'fullstack-smoke', status: 'queued', time: 'queued' },
-              { name: 'coverage-gate', status: 'success', time: '24s' },
-            ].map((r) => (
-              <div key={r.name} className="px-2 py-1" style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION, backgroundColor: 'rgba(30,41,59,0.5)', borderRadius: PREV_XS }}>
-                <span style={ps.dot(r.status === 'running' ? ps.colors.info : r.status === 'success' ? ps.colors.healthy : PREV_CLR_MUTED)} />
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_3}>{r.name}</span>
-                <span style={{ color: PREV_CLR_MUTED, fontSize: PREV_FS_MICRO }}>{r.time}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'recent_failures':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.error)} /> Recent Failures</div>
-          <div style={ps.col}>
-            {[
-              { wf: 'nightly-test-suite', step: 'e2e-tests', ago: '2h ago' },
-              { wf: 'build', step: 'lint', ago: '5h ago' },
-            ].map((f) => (
-              <div key={f.wf} className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(239,68,68,0.1)', borderRadius: PREV_XS, borderLeft: `${PREV_BORDER_THIN} solid ${ps.colors.error}` }}>
-                <div style={WIDGET_EXPORT_MODAL_DIV_STYLE_2}>
-                  <div style={{ fontSize: PREV_FS_CAPTION, fontWeight: 600, color: ps.colors.error }}>{f.wf}</div>
-                  <div style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED }}>Failed at: {f.step}</div>
-                </div>
-                <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED }}>{f.ago}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'issue_activity_chart':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Daily Issues &amp; PRs</div>
-          <div style={{ display: 'flex', alignItems: 'flex-end', gap: PREV_BAR_GAP, height: '60px', marginBottom: PREV_SM }}>
-            {[4, 7, 3, 8, 5, 6, 9, 2, 5, 7, 4, 6, 3].map((v, i) => (
-              <div key={i} style={{ flex: 1, display: 'flex', flexDirection: 'column', gap: PREV_HAIRLINE_GAP, justifyContent: 'flex-end', height: '100%' }}>
-                <div style={{ height: `${v * PREV_BAR_OPENED_SCALE}px`, backgroundColor: ps.colors.info, borderRadius: '1px', opacity: 0.7 }} />
-                <div style={{ height: `${Math.max(0, (PREV_BAR_CLOSED_BASE - v) * PREV_BAR_CLOSED_SCALE)}px`, backgroundColor: ps.colors.healthy, borderRadius: '1px', opacity: 0.5 }} />
-              </div>
-            ))}
-          </div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY, color: ps.colors.info }}>23</span><span style={ps.statLbl}>Opened</span></div>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY, color: ps.colors.healthy }}>18</span><span style={ps.statLbl}>Closed</span></div>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY, color: ps.colors.purple }}>12</span><span style={ps.statLbl}>Merged</span></div>
-          </div>
-        </div>
-      )
-
-    case 'github_ci_monitor':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> GitHub CI Monitor</div>
-          <div style={{ textAlign: 'center', marginBottom: PREV_SM }}>
-            <div style={{ fontSize: PREV_FS_HEADLINE, fontWeight: 700, color: ps.colors.healthy }}>94%</div>
-            <div style={ps.muted}>Pass rate (7d)</div>
-          </div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY }}>156</span><span style={ps.statLbl}>Runs</span></div>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY, color: ps.colors.healthy }}>147</span><span style={ps.statLbl}>Passed</span></div>
-            <div style={ps.statBlock}><span style={{ ...ps.statVal, fontSize: PREV_FS_BODY, color: ps.colors.error }}>9</span><span style={ps.statLbl}>Failed</span></div>
-          </div>
-        </div>
-      )
-
-    case 'github_activity':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> GitHub Activity</div>
-          <div style={ps.col}>
-            {[
-              { label: 'PRs merged (7d)', value: '24', color: ps.colors.purple },
-              { label: 'Issues opened', value: '8', color: ps.colors.info },
-              { label: 'Contributors', value: '6', color: ps.colors.healthy },
-              { label: 'Latest release', value: 'v0.3.22', color: PREV_CLR_SECONDARY },
-            ].map((item) => (
-              <div key={item.label} style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={{ color: PREV_CLR_MUTED }}>{item.label}</span>
-                <span style={{ fontWeight: 600, color: item.color }}>{item.value}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'cluster_metrics':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Cluster Metrics</div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: PREV_CLR_CPU }}>62%</span>
-              <span style={ps.statLbl}><TechnicalAcronym term="CPU">CPU</TechnicalAcronym></span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: PREV_CLR_MEM }}>78%</span>
-              <span style={ps.statLbl}>Memory</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT_SM, color: ps.colors.healthy }}>45</span>
-              <span style={ps.statLbl}>Pods</span>
-            </div>
-          </div>
-          <div style={{ display: 'flex', alignItems: 'flex-end', gap: PREV_DOTS_GAP, height: '40px', marginTop: PREV_SM }}>
-            {[40, 55, 62, 58, 70, 65, 72, 68, 75, 62].map((v, i) => (
-              <div key={i} style={{ flex: 1, height: `${v * 0.55}px`, backgroundColor: ps.colors.info, borderRadius: '1px', opacity: 0.6 }} />
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'workload_status':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Workload Status</div>
-          <div style={ps.col}>
-            {[
-              { name: 'Deployments', running: 12, total: 14, color: ps.colors.healthy },
-              { name: 'StatefulSets', running: 3, total: 3, color: ps.colors.healthy },
-              { name: 'DaemonSets', running: 4, total: 5, color: ps.colors.warning },
-            ].map((w) => (
-              <div key={w.name} className="px-2 py-1" style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_2}>{w.name}</span>
-                <span style={{ color: w.running === w.total ? ps.colors.healthy : ps.colors.warning }}>
-                  {w.running}/{w.total} ready
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'app_status':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> Application Status</div>
-          <div style={ps.col}>
-            {[
-              { name: 'frontend', clusters: 3, status: 'healthy' },
-              { name: 'api-gateway', clusters: 2, status: 'healthy' },
-              { name: 'worker', clusters: 2, status: 'degraded' },
-              { name: 'scheduler', clusters: 1, status: 'healthy' },
-            ].map((a) => (
-              <div key={a.name} style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_2}>{a.name}</span>
-                <span style={{ color: PREV_CLR_MUTED, fontSize: PREV_FS_MICRO }}>{a.clusters} clusters</span>
-                <span style={ps.dot(a.status === 'healthy' ? ps.colors.healthy : ps.colors.warning)} />
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'namespace_overview':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.info)} /> Namespace Overview</div>
-          <div style={ps.col}>
-            {[
-              { ns: 'default', pods: 12, deploys: 4 },
-              { ns: 'kube-system', pods: 24, deploys: 8 },
-              { ns: 'monitoring', pods: 6, deploys: 3 },
-              { ns: 'production', pods: 18, deploys: 6 },
-            ].map((n) => (
-              <div key={n.ns} className="px-2 py-1" style={{ ...ps.row, justifyContent: 'space-between', fontSize: PREV_FS_CAPTION }}>
-                <span style={{ fontWeight: 500, color: PREV_CLR_CPU }}>{n.ns}</span>
-                <span style={{ color: PREV_CLR_MUTED }}>{n.pods} pods</span>
-                <span style={{ color: PREV_CLR_MUTED }}>{n.deploys} deploys</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'console_ai_health_check':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.healthy)} /> AI Health Check</div>
-          <div style={{ textAlign: 'center', marginBottom: PREV_SM }}>
-            <div style={{ fontSize: PREV_FS_FEATURED, fontWeight: 700, color: ps.colors.healthy }}>Healthy</div>
-            <div style={ps.muted}>AI analysis complete</div>
-          </div>
-          <div style={ps.col}>
-            {[
-              { finding: 'All nodes responding', severity: 'ok' },
-              { finding: 'Pod restart rate normal', severity: 'ok' },
-              { finding: 'Memory pressure on worker-2', severity: 'warn' },
-            ].map((f) => (
-              <div key={f.finding} className="px-2 py-1" style={{ ...ps.row, fontSize: PREV_FS_CAPTION }}>
-                <span style={ps.dot(f.severity === 'ok' ? ps.colors.healthy : ps.colors.warning)} />
-                <span style={{ color: f.severity === 'ok' ? PREV_CLR_SECONDARY : ps.colors.warning }}>{f.finding}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )
-
-    case 'console_ai_offline_detection':
-      return (
-        <div style={ps.card}>
-          <div style={ps.title}><span style={ps.dot(ps.colors.warning)} /> AI Offline Detection</div>
-          <div style={ps.row}>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.healthy }}>11</span>
-              <span style={ps.statLbl}>Online</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.error }}>1</span>
-              <span style={ps.statLbl}>Offline</span>
-            </div>
-            <div style={ps.statBlock}>
-              <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: ps.colors.warning }}>2</span>
-              <span style={ps.statLbl}>GPUs down</span>
-            </div>
-          </div>
-          <div className="px-2 py-1" style={{ ...ps.row, backgroundColor: 'rgba(239,68,68,0.1)', borderRadius: PREV_XS, borderLeft: `${PREV_BORDER_THIN} solid ${ps.colors.error}`, marginTop: PREV_SM }}>
-            <span style={{ fontSize: PREV_FS_CAPTION, color: ps.colors.error, fontWeight: 600 }}>worker-4</span>
-            <span style={{ fontSize: PREV_FS_MICRO, color: PREV_CLR_MUTED, marginLeft: 'auto' }}>unreachable 12m</span>
-          </div>
-        </div>
-      )
-
-    default:
-      return <GenericCardPreview card={card} />
-  }
+  if (CARD_PREVIEW_A_TYPES.has(cardType)) return <CardPreviewA cardType={cardType} card={card} />
+  return <CardPreviewB cardType={cardType} card={card} />
 }
 
-// Generic card preview based on category
-function GenericCardPreview({ card }: { card: WidgetCardDefinition }) {
-  const categoryData: Record<string, { dot: string; items: { label: string; value: string; color?: string }[] }> = {
-    cluster: { dot: ps.colors.healthy, items: [{ label: 'Ready', value: '3/4', color: ps.colors.healthy }, { label: 'Nodes', value: '12' }, { label: 'Version', value: 'v1.28' }] },
-    workload: { dot: ps.colors.info, items: [{ label: 'Running', value: '45', color: ps.colors.healthy }, { label: 'Pending', value: '2', color: ps.colors.warning }, { label: 'Failed', value: '1', color: ps.colors.error }] },
-    gpu: { dot: ps.colors.purple, items: [{ label: 'Total', value: '32' }, { label: 'Allocated', value: '24', color: ps.colors.purple }, { label: 'Available', value: '8', color: ps.colors.healthy }] },
-    security: { dot: ps.colors.warning, items: [{ label: 'Critical', value: '2', color: ps.colors.error }, { label: 'Warning', value: '5', color: ps.colors.warning }, { label: 'Info', value: '8', color: ps.colors.info }] },
-    monitoring: { dot: ps.colors.info, items: [{ label: 'Active', value: '3', color: ps.colors.info }, { label: 'Resolved', value: '12', color: ps.colors.healthy }, { label: 'Silenced', value: '1' }] },
-    'ci-cd': { dot: ps.colors.info, items: [{ label: 'Runs', value: '36', color: ps.colors.info }, { label: 'Passed', value: '34', color: ps.colors.healthy }, { label: 'Failed', value: '2', color: ps.colors.error }] } }
-  const data = categoryData[card.category] || categoryData.monitoring
-  return (
-    <div style={ps.card}>
-      <div style={ps.title}><span style={ps.dot(data.dot)} /> {card.displayName}</div>
-      <div style={ps.row}>
-        {data.items.map((item) => (
-          <div key={item.label} style={ps.statBlock}>
-            <span style={{ ...ps.statVal, fontSize: PREV_FS_STAT, color: item.color || '#f9fafb' }}>{item.value}</span>
-            <span style={ps.statLbl}>{item.label}</span>
-          </div>
-        ))}
-      </div>
-    </div>
-  )
-}
-
-// --- Stat previews ---
 function StatPreview({ statIds }: { statIds: string[] }) {
   return (
     <div style={{ ...ps.card, display: 'flex', flexWrap: 'wrap', gap: PREV_SM, padding: PREV_CARD_PAD, overflow: 'hidden' }}>
@@ -821,7 +95,6 @@ function StatPreview({ statIds }: { statIds: string[] }) {
   )
 }
 
-// --- Template previews ---
 function TemplatePreview({ templateId }: { templateId: string }) {
   const template = WIDGET_TEMPLATES[templateId]
   if (!template) return null
@@ -879,61 +152,27 @@ function TemplatePreview({ templateId }: { templateId: string }) {
   )
 }
 
-// Nightly E2E preview with sample status dots
-function NightlyE2EPreview() {
-  const platforms = [
-    {
-      name: 'OCP', color: '#f97316',
-      guides: [
-        { acronym: 'IS', dots: ['g','g','r','g','g','g','g'] },
-        { acronym: 'PD', dots: ['g','g','g','g','g','g','g'] },
-        { acronym: 'PPC', dots: ['g','r','g','g','g','r','g'] },
-        { acronym: 'TPC', dots: ['g','g','g','r','g','g','g'] },
-        { acronym: 'WEP', dots: ['g','g','g','g','g','g','b'] },
-        { acronym: 'WVA', dots: ['g','r','g','g','r','g','g'] },
-      ] },
-    {
-      name: 'GKE', color: '#3b82f6',
-      guides: [
-        { acronym: 'IS', dots: ['g','g','g','g','g','g','g'] },
-        { acronym: 'PD', dots: ['r','g','g','g','g','g','g'] },
-        { acronym: 'WEP', dots: ['g','g','g','g','g','g','g'] },
-      ] },
-    {
-      name: 'CKS', color: '#a855f7',
-      guides: [
-        { acronym: 'IS', dots: [] as string[] },
-        { acronym: 'PD', dots: [] as string[] },
-        { acronym: 'WEP', dots: [] as string[] },
-      ] },
-  ]
-  const dotColor: Record<string, string> = { g: '#22c55e', r: '#ef4444', b: '#60a5fa' }
-
-  return (
-    <div style={{ ...ps.card, width: 320, fontSize: PREV_FS_CAPTION, padding: PREV_CARD_PAD }}>
-      <div style={ps.title}><span style={ps.dot('#22c55e')} /> Nightly E2E Status</div>
-      <div style={{ display: 'flex', gap: PREV_LG, marginBottom: PREV_SM }}>
-        <div><span style={{ fontSize: PREV_FS_STAT, fontWeight: 700, color: '#a855f7' }}>87%</span><div style={ps.muted}>Pass Rate</div></div>
-        <div><span style={{ fontSize: PREV_FS_STAT, fontWeight: 700 }}>16</span><div style={ps.muted}>Guides</div></div>
-        <div><span style={{ fontSize: PREV_FS_STAT, fontWeight: 700, color: '#ef4444' }}>3</span><div style={ps.muted}>Failing</div></div>
+export function WidgetPreview({ config }: { config: WidgetConfig | null }) {
+  if (!config) {
+    return (
+      <div className="text-center text-muted-foreground">
+        <Smartphone className="w-12 h-12 mx-auto mb-2 opacity-50" />
+        <p className="text-sm">Select an item to preview</p>
       </div>
-      {platforms.map((p) => (
-        <div key={p.name} className="mb-1">
-          <div className="mb-1" style={{ color: p.color, fontWeight: 600, fontSize: PREV_FS_MICRO }}>{p.name}</div>
-          {p.guides.map((g) => (
-            <div key={`${p.name}-${g.acronym}`} className="flex items-center gap-1" style={{ marginBottom: PREV_HAIRLINE_GAP }}>
-              <span style={WIDGET_EXPORT_MODAL_SPAN_STYLE_4}>{g.acronym}</span>
-              <div style={{ display: 'flex', gap: PREV_DOTS_GAP }}>
-                {g.dots.length > 0 ? g.dots.map((d, i) => (
-                  <span key={i} style={{ width: 6, height: 6, borderRadius: '50%', backgroundColor: dotColor[d], display: 'inline-block', ...(d === 'b' ? { animation: 'pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite' } : {}) }} />
-                )) : (
-                  <span style={{ color: PREV_CLR_MUTED, fontSize: PREV_FS_LABEL }}>no runs</span>
-                )}
-              </div>
-            </div>
-          ))}
-        </div>
-      ))}
-    </div>
-  )
+    )
+  }
+
+  if (config.type === 'card' && config.cardType) {
+    return <CardPreview cardType={config.cardType} />
+  }
+
+  if (config.type === 'stat' && config.statIds) {
+    return <StatPreview statIds={config.statIds} />
+  }
+
+  if (config.type === 'template' && config.templateId) {
+    return <TemplatePreview templateId={config.templateId} />
+  }
+
+  return null
 }
